Fix model type mismatches for JSON deserialization
- IssueStats: DaysOpen, DaysSinceActivity changed int → double
- PrTriageStats: DaysOpen changed int → double
- NuGetStats: VersionStats.Downloads changed long → long?

JSON returns floats (5.0) for day counts and null for missing
download data. These type mismatches caused silent deserialization
failures.

+++ src/Dashboard/Services/PrTriageStats.cs
@@ -25,7 +25,7 @@ public record TriagedPullRequest(
     string AuthorType,
     DateTime CreatedAt,
     DateTime UpdatedAt,
-    int DaysOpen,
+    double DaysOpen,
     string AgeCategory,
     int FilesChanged,
     int Additions,
